chore: rearrange utils based on context
